feat: add lurkers force flag

+++ apps/bots/internal/chat_client/handlers_remove_lurker.go
@@ -18,7 +18,7 @@ func (c *ChatClient) removeUserFromLurkers(userId string) {
 		return
 	}
 
-	if ignoredUser.ID != "" {
+	if ignoredUser.ID != "" && !ignoredUser.Force {
 		err = c.services.DB.Delete(ignoredUser).Error
 		if err != nil {
 			c.services.Logger.Error(
